automation_project_API/requirements.txt
pytest
pyyaml
allure-pytest
